fix: remove post layout
Fixes an issue where navigating to the a post page at `/posts/[slug]` might cause the page to load scrolled part of the way down. Apparently, layouts wrap their children in a div with a data attribute that next looks for to determine where to scroll to on load (or render?). Layouts are also not re-rendered. I assume this means that next couldn't figure out where the right scroll position should be if you navigate to post, go back, and select a different post since all post pages shared a nested layout.

GH issue comments for reference:

https://github.com/vercel/next.js/issues/42492#issuecomment-1320993510
https://github.com/vercel/next.js/issues/42492#issuecomment-1345544864